refactor(queries): consolidate query files into a single queries directory
Moved all query files from the root `lib` directory into a new `queries` subdirectory. Updated all import paths to reflect the new structure. This change improves code organization and maintainability by grouping related query files together.
