+++ app/(app)/page.tsx
@@ -1,5 +1,5 @@
 import { getSanityDocuments } from '@/sanity/lib/fetch-sanity-data';
-import { HOME_PAGE_QUERY } from '@/sanity/lib/page.query';
+import { HOME_PAGE_QUERY } from '@/sanity/lib/queries/page.query';
 import { PageBuilder } from '@/components/PageBuilder';
 
 export default async function Home() {
