+++ sanity/lib/queries/category.query.ts
