+++ sanity/lib/queries/project.query.ts
